my projict/ecommerce_marketplace/requirements.txt
Flask
pymongo==3.12.0
Flask-Login==0.5.0
Flask-WTF==0.15.1
Werkzeug==2.0.1
python-dotenv==0.19.0
Pillow
bcrypt==3.2.0
email-validator==1.1.3 
torch 
torchvision 
faiss-cpu
open_clip_torch
